add:firebase.ts&&store.ts(redux)
