--- firebase.ts
@@ -1,17 +1,0 @@
-import { initializeApp } from "firebase/app"
-import { getFirestore } from 'firebase/firestore';
-
-const firebaseConfig = {
-  apiKey: process.env.FIREBASE_API_KEY,
-  authDomain: process.env.FIREBASE_AUTH_DOMAIN,
-  projectId: process.env.FIREBASE_PROJECT_ID,
-  storageBucket: process.env.FIREBASE_STORAGE_BUCKET,
-  appId: process.env.FIREBASE_APP_ID,
-}
-
-export const app = initializeApp(firebaseConfig)
-
-
-
-export const db = getFirestore(app);
-
